Move copy button to separate file
